Fixed issue where manager could update another manager's records

+++ quiz_application/manager.py
@@ -22,7 +22,7 @@ def create_subject():
     if request.method == 'POST':
         subject_name = request.form.get('name')
         about = request.form.get('about')
-        new_subject = Subject (name=subject_name, about=about)
+        new_subject = Subject (name=subject_name, about=about, manager_id=current_user.id)
         db.session.add(new_subject)
         db.session.commit()
         flash("Success!", category='success') 
@@ -34,31 +34,38 @@ def create_subject():
 @is_manager
 def update_subject(id): 
     subject_to_update = Subject.query.get_or_404(id)
-    if request.method == 'POST':
-        subject_to_update.name = request.form.get('name')
-        subject_to_update.about = request.form.get('about')
-        try:
-            db.session.commit()
-            flash("Subject updated successfully!", category='success')
-            return redirect('/manager/manage-subjects')
-        except:
-            flash("Couldn't update subject", category="error") 
-        return redirect('/manager/manage-subjects')
+    if subject_to_update.manager_id == current_user.id:
+        if request.method == 'POST':
+            subject_to_update.name = request.form.get('name')
+            subject_to_update.about = request.form.get('about')
+            try:
+                db.session.commit()
+                flash("Subject updated successfully!", category='success')
+                return redirect('/manager/manage-subjects')
+            except:
+                flash("Couldn't update subject", category="error")  
+    else:
+        flash("You can only update your subject", category="error")
+        return redirect("/manager/") 
     return render_template('/manager/update-subject.html', user=current_user, subject=subject_to_update)
      
 @manager.route('delete-subject/<int:id>')
 @login_required
 @is_manager
-def delete_subject(id): 
-    try:
-        subject_to_delete = Subject.query.get_or_404(id)
-        db.session.delete(subject_to_delete)
-        db.session.commit()
-        flash("Successfully deleted!", category="success")
+def delete_subject(id):
+    subject_to_delete = Subject.query.get_or_404(id)
+    if subject_to_delete.manager_id == current_user.id:
+        try:
+            db.session.delete(subject_to_delete)
+            db.session.commit()
+            flash("Successfully deleted!", category="success")
+            return redirect("/manager/manage-subjects")
+        except:
+            flash("Something went wrong!", category="error")
+            return redirect("manager/manage-subjects")
+    else:
+        flash("You can only delete your subjects", category="error")
         return redirect("/manager/manage-subjects")
-    except:
-        flash("Something went wrong!", category="error")
-        return redirect("manager/manage-subjects")
 
 @manager.route('/manage-questions')
 @login_required
@@ -79,10 +86,10 @@ def create_question():
         choice_two = request.form['choice_two']
         choice_three = request.form['choice_three']
         answer = request.form['correct_answer']
-        new_question = Question(question=question, subject_id=subject_id, choice_one=choice_one, choice_two=choice_two, choice_three=choice_three, answer=answer)
+        new_question = Question(question=question, subject_id=subject_id, choice_one=choice_one, choice_two=choice_two, choice_three=choice_three, answer=answer, manager_id=current_user.id)
         db.session.add(new_question)
         db.session.commit()
-        flash("Successfuly created the questoin!", category='success') 
+        flash("Successfuly created the question!", category='success') 
         return redirect('/manager/manage-questions')
     return render_template("manager/create-question.html", subjects=subjects, user=current_user)
 
@@ -92,33 +99,41 @@ def create_question():
 def update_question(id):
     subjects = Subject.query.all()
     question_to_update = Question.query.get_or_404(id)
-    if request.method == "POST":
-        question_to_update.question = request.form['question']
-        question_to_update.subject_id = request.form['subject_id']
-        question_to_update.choice_one = request.form['choice_one']
-        question_to_update.choice_two = request.form['choice_two']
-        question_to_update.choice_three = request.form['choice_three']
-        question_to_update.answer = request.form['correct_answer']
-        try: 
-            db.session.commit()
-            flash("Successfully updated question!", category='success') 
-            return redirect('/manager/manage-questions')
-        except:
-            flash("something went wrong!", category="error")
+    if question_to_update.manager_id == current_user.id:
+        if request.method == "POST":
+            question_to_update.question = request.form['question']
+            question_to_update.subject_id = request.form['subject_id']
+            question_to_update.choice_one = request.form['choice_one']
+            question_to_update.choice_two = request.form['choice_two']
+            question_to_update.choice_three = request.form['choice_three']
+            question_to_update.answer = request.form['correct_answer']
+            try: 
+                db.session.commit()
+                flash("Successfully updated question!", category='success') 
+                return redirect('/manager/manage-questions')
+            except:
+                flash("something went wrong!", category="error")
+    else:
+        flash("You can only update your question!", category="error")
+        return redirect("/manager/manage-questions")
     return render_template("manager/update-question.html", question=question_to_update , subjects=subjects, user=current_user)
 
 @manager.route("delete-question/<int:id>")
 @login_required
 @is_manager
 def delete_question(id):
-    try:
-        question_to_delete = Question.query.get_or_404(id)
-        db.session.delete(question_to_delete)
-        db.session.commit()
-        flash("Successfully deleted!", category="success")
-        return redirect('/manager/manage-questions')
-    except:
-        flash("Something went wrong!", category="error")
+    question_to_delete = Question.query.get_or_404(id)
+    if question_to_delete.manager_id == current_user.id: 
+        try:
+            db.session.delete(question_to_delete)
+            db.session.commit()
+            flash("Successfully deleted!", category="success")
+            return redirect('/manager/manage-questions')
+        except:
+            flash("Something went wrong!", category="error")
+    else:
+        flash("You can only delete your questions", category="error")
+        return redirect("/manager/manage-questions")
 
 @manager.route("/students-list")
 @login_required
